+++ Acr.Ble.Uwp/Device.cs
@@ -42,10 +42,9 @@ namespace Acr.Ble
 
         public IObservable<object> Connect(GattConnectionConfig config)
         {
-            config = config ?? GattConnectionConfig.DefaultConfiguration;
-
             // TODO: config auto reconnect?
             // TODO: monitor devicewatcher - if removed d/c, if added AND paired - connected
+            this.native.
             return Observable.Create<object>(ob =>
             {
                 if (this.Status == ConnectionStatus.Connected)